pyproject.toml
[tool.poetry]
name = "licensor"
version = "0.1.0"
description = ""
authors = ["Ihor Savenko <csigorek@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.115.6"
uvicorn = "^0.34.0"
pydantic = "^2.10.4"
python-jose = "^3.3.0"
requests = "^2.32.3"
psutil = "^6.1.1"
uuid = "^1.30"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
